Add unit tests.

+++ src/Authentication/Authentication/TokenCache/WindowsTokenCache.cs
@@ -14,38 +14,52 @@ namespace Microsoft.Graph.PowerShell.Authentication.TokenCache
         /// <summary>
         /// Gets a decrypted token from the token cache.
         /// </summary>
-        /// <param name="tokenCacheFilePath">Path to the token cache file to read from.</param>
+        /// <param name="appId">The app/client id..</param>
         /// <returns>A decrypted access token.</returns>
-        public static byte[] GetToken(string tokenCacheFilePath)
+        public static byte[] GetToken(string appId)
         {
             if (!Directory.Exists(Constants.TokenCacheDirectory))
                 Directory.CreateDirectory(Constants.TokenCacheDirectory);
 
-            return File.Exists(tokenCacheFilePath)
-                        ? ProtectedData.Unprotect(File.ReadAllBytes(tokenCacheFilePath), null, DataProtectionScope.CurrentUser)
-                        : new byte[0];
+            string tokenCacheFilePath = Path.Combine(Constants.TokenCacheDirectory, $"{appId}cache.bin3");
+
+            return File.Exists(tokenCacheFilePath) ?
+                ProtectedData.Unprotect(
+                    encryptedData: File.ReadAllBytes(tokenCacheFilePath),
+                    optionalEntropy: null,
+                    scope: DataProtectionScope.CurrentUser)
+                : new byte[0];
         }
 
         /// <summary>
         /// Sets an encrypted access token to the token cache.
         /// </summary>
-        /// <param name="tokenCacheFilePath">Path to the token cache file path to write to.</param>
+        /// <param name="appId">The app/client id..</param>
         /// <param name="plainContent">Plain access token to securely write to the token cache file.</param>
-        public static void SetToken(string tokenCacheFilePath, byte[] plainContent)
+        public static void SetToken(string appId, byte[] plainContent)
         {
             if (!Directory.Exists(Constants.TokenCacheDirectory))
                 Directory.CreateDirectory(Constants.TokenCacheDirectory);
 
-            File.WriteAllBytes(tokenCacheFilePath, ProtectedData.Protect(plainContent, null, DataProtectionScope.CurrentUser));
+            string tokenCacheFilePath = Path.Combine(Constants.TokenCacheDirectory, $"{appId}cache.bin3");
+
+            File.WriteAllBytes(tokenCacheFilePath,
+                ProtectedData.Protect(
+                    userData: plainContent,
+                    optionalEntropy: null,
+                    scope: DataProtectionScope.CurrentUser));
         }
 
         /// <summary>
         /// Deletes an access token cache file/
         /// </summary>
-        /// <param name="tokenCacheFilePath">Token cache file path to delete.</param>
-        public static void DeleteToken(string tokenCacheFilePath)
+        /// <param name="appId">The app/client id to delete its token cache file.</param>
+        public static void DeleteToken(string appId)
         {
-            File.Delete(tokenCacheFilePath);
+            string tokenCacheFilePath = Path.Combine(Constants.TokenCacheDirectory, $"{appId}cache.bin3");
+
+            if (File.Exists(tokenCacheFilePath))
+                File.Delete(tokenCacheFilePath);
         }
     }
 }
